[ch] changed logic implementation in route to controllers
