+++ server/test/productsTest.js
@@ -91,6 +91,54 @@ describe('Get A Product', () => {
 });
 
 describe('Create New Product', () => {
+  it('return unauthorized because user is not admin', (done) => {
+    chai.request(app).post('/api/v1/users/login')
+      .send({
+        email: 'example2@gmail.com', password: '123456',
+      })
+      .end((err, res) => {
+        const { token } = res.body;
+        expect(res).to.have.status(200);
+        expect(res.body).to.be.an('object');
+        expect(res.body.success).to.equal(true);
+        chai.request(app).post('/api/v1/products')
+          .set('Authorization', token)
+          .end((error, data) => {
+            expect(data).to.have.status(401);
+            expect(data.body.message).to.equal('Unauthorized');
+            done();
+          });
+      });
+  });
+
+  it('return validation error if no data is sent', (done) => {
+    chai.request(app).post('/api/v1/users/login')
+      .send({
+        email: 'example@gmail.com', password: '123456',
+      })
+      .end((err, res) => {
+        const { token } = res.body;
+        expect(res).to.have.status(200);
+        expect(res.body).to.be.an('object');
+        expect(res.body.success).to.equal(true);
+        chai.request(app).post('/api/v1/products')
+          .set('Authorization', token)
+          .send({
+            name: 'iPhone',
+          })
+          .end((error, data) => {
+            expect(data).to.have.status(400);
+            expect(data.body).to.be.an('object');
+            // expect(data.body.name).to.equal('Name field is required');
+            expect(data.body.description).to.equal('Description field is required');
+            expect(data.body.price).to.equal('Price field is required');
+            expect(data.body.quantity).to.equal('Quantity field is required');
+            done();
+          });
+      });
+  });
+
+
   it('create a new product', (done) => {
     chai.request(app).post('/api/v1/users/login')
       .send({
@@ -110,30 +158,6 @@ describe('Create New Product', () => {
             expect(data).to.have.status(201);
             expect(data.body).to.be.an('object');
             expect(data.body.message).to.equal('Product added successfully');
-            done();
-          });
-      });
-  });
-
-  it('return validation error if no data is sent', (done) => {
-    chai.request(app).post('/api/v1/users/login')
-      .send({
-        email: 'example@gmail.com', password: '123456',
-      })
-      .end((err, res) => {
-        const { token } = res.body;
-        expect(res).to.have.status(200);
-        expect(res.body).to.be.an('object');
-        expect(res.body.success).to.equal(true);
-        chai.request(app).post('/api/v1/products')
-          .set('Authorization', token)
-          .end((error, data) => {
-            expect(data).to.have.status(400);
-            expect(data.body).to.be.an('object');
-            expect(data.body.name).to.equal('Name field is required');
-            expect(data.body.description).to.equal('Description field is required');
-            expect(data.body.price).to.equal('Price field is required');
-            expect(data.body.quantity).to.equal('Quantity field is required');
             done();
           });
       });
@@ -210,6 +234,33 @@ describe('Delete A Product', () => {
             expect(data.body).to.be.an('object');
             expect(data.body.message).to.equal('Product with id 59 not found.');
             done();
+          });
+      });
+  });
+
+  it('should return product not found because the product has just been deleted', (done) => {
+    chai.request(app).post('/api/v1/users/login')
+      .send({
+        email: 'example@gmail.com', password: '123456',
+      })
+      .end((err, res) => {
+        const { token } = res.body;
+        expect(res).to.have.status(200);
+        expect(res.body).to.be.an('object');
+        expect(res.body.success).to.equal(true);
+        chai.request(app).del('/api/v1/products/2')
+          .set('Authorization', token)
+          .end((error, data) => {
+            expect(data).to.have.status(200);
+            expect(data.body).to.be.an('object');
+            chai.request(app).del('/api/v1/products/2')
+              .set('Authorization', token)
+              .end((error2, data2) => {
+                expect(data2).to.have.status(400);
+                expect(data2.body).to.be.an('object');
+                expect(data2.body.message).to.equal('Product with id 2 not found.');
+                done();
+              });
           });
       });
   });
