+++ server/test/userTest.js
@@ -6,50 +6,105 @@ const { expect } = chai;
 
 chai.use(chaiHttp);
 
-describe('Register Route', () => {
-  it('create a user and return user details', (done) => {
-    chai.request(app).post('/api/v1/users/register')
+describe('Signup Route', () => {
+  // it('create a user and return user details', (done) => {
+  //   chai.request(app).post('/api/v1/users/signup')
+  //     .send({
+  //       email: 'a@gmail.com',
+  //       name: 'John Example',
+  //       password: '123456',
+  //       type: '1',
+  //     })
+  //     .end((err, res) => {
+  //       expect(res).to.have.status(201);
+  //       expect(res.body).to.be.an('object');
+  //       expect(res.body.message).to.equal('User Created Successfully');
+  //       done();
+  //     });
+  // });
+
+  it('create a user and return user details', (done) => { 
+    chai.request(app).post('/api/v1/users/login')
       .send({
-        email: 'a@gmail.com',
-        name: 'John Example',
+        email: 'example@gmail.com',
         password: '123456',
-        type: '1',
       })
-      .end((err, res) => {
-        expect(res).to.have.status(201);
-        expect(res.body).to.be.an('object');
-        expect(res.body.message).to.equal('User Created Successfully');
-        done();
+      .end((loginerr, loginres) => {
+        const { token } = loginres.body;
+        expect(loginres).to.have.status(200);
+        expect(loginres.body).to.be.an('object');
+        expect(loginres.body.success).to.equal(true);
+        expect(loginres.body.token).to.include('Bearer');
+
+        chai.request(app).post('/api/v1/users/signup')
+          .set('Authorization', token)
+          .send({
+            email: 'a@gmail.com',
+            name: 'John Example',
+            password: '123456',
+            type: '1',
+          })
+          .end((err, res) => {
+            expect(res).to.have.status(201);
+            expect(res.body).to.be.an('object');
+            expect(res.body.message).to.equal('User Created Successfully');
+            done();
+          });
       });
   });
 
   it('return validation error if no data is sent', (done) => {
-    chai.request(app).post('/api/v1/users/register')
-      .end((err, res) => {
-        expect(res).to.have.status(400);
-        expect(res.body).to.be.an('object');
-        expect(res.body.email).to.equal('Email field is required');
-        expect(res.body.password).to.equal('Password field is required');
-        expect(res.body.name).to.equal('Name field is required');
-        expect(res.body.type).to.equal('Type field is required');
-
-        done();
+    chai.request(app).post('/api/v1/users/login')
+      .send({
+        email: 'example@gmail.com',
+        password: '123456',
+      })
+      .end((loginerr, loginres) => {
+        const { token } = loginres.body;
+        expect(loginres).to.have.status(200);
+        expect(loginres.body).to.be.an('object');
+        expect(loginres.body.success).to.equal(true);
+        expect(loginres.body.token).to.include('Bearer');
+        chai.request(app).post('/api/v1/users/signup')
+          .set('Authorization', token)
+          .end((err, res) => {
+            expect(res).to.have.status(400);
+            expect(res.body).to.be.an('object');
+            expect(res.body.email).to.equal('Email field is required');
+            expect(res.body.password).to.equal('Password field is required');
+            expect(res.body.name).to.equal('Name field is required');
+            expect(res.body.type).to.equal('Type field is required');
+            done();
+          });
       });
   });
 
   it('return email already exist', (done) => {
-    chai.request(app).post('/api/v1/users/register')
+    chai.request(app).post('/api/v1/users/login')
       .send({
         email: 'example@gmail.com',
-        name: 'John Example',
         password: '123456',
-        type: '1',
       })
-      .end((err, res) => {
-        expect(res).to.have.status(409);
-        expect(res.body).to.be.an('object');
-        expect(res.body.email).to.equal('Email Already Exist');
-        done();
+      .end((loginerr, loginres) => {
+        const { token } = loginres.body;
+        expect(loginres).to.have.status(200);
+        expect(loginres.body).to.be.an('object');
+        expect(loginres.body.success).to.equal(true);
+        expect(loginres.body.token).to.include('Bearer');
+        chai.request(app).post('/api/v1/users/signup')
+          .set('Authorization', token)
+          .send({
+            email: 'example@gmail.com',
+            name: 'John Example',
+            password: '123456',
+            type: '1',
+          })
+          .end((err, res) => {
+            expect(res).to.have.status(409);
+            expect(res.body).to.be.an('object');
+            expect(res.body.email).to.equal('Email Already Exist');
+            done();
+          });
       });
   });
 });
