+++ server/test/usersValidationTest.js
@@ -27,9 +27,9 @@ describe('Login Validation', () => {
 });
 
 
-describe('Register Validation', () => {
+describe('Login Validation', () => {
   it('returns empty object because all validation is passed', (done) => {
-    const result = usersValidation.validateRegisterInput(
+    const result = usersValidation.validateLoginInput(
       {
         email: 'example@gmail.com', password: '123456', name: 'Example', type: '2',
       },
@@ -40,13 +40,11 @@ describe('Register Validation', () => {
   });
 
   it('returns object of validation required', (done) => {
-    const result = usersValidation.validateRegisterInput({});
+    const result = usersValidation.validateLoginInput({});
     expect(result.isValid).to.equal(false);
     expect(Object.keys(result.errors).length).to.be.greaterThan(0);
     expect(result.errors.email).to.equal('Email field is required');
     expect(result.errors.password).to.equal('Password field is required');
-    expect(result.errors.name).to.equal('Name field is required');
-    expect(result.errors.type).to.equal('Type field is required');
     expect(result.errors).to.be.an('object');
     done();
   });
